Separated fns related to generating schema into different modules. Generating of random data for tables complete
