+++ db/schemaGenerator/words.js
@@ -1,0 +1,4 @@
+var words = "culture fly style essentials legendary global icon premium gear swoosh ignite virgil clothing vapormax exclusive flight pattern run tech kickoff world homage air explore collection knit reserved limitless shilouette innovation access versatile member sport equality offwhite ethos attack zoom stealth quick force ultra womens flyknit upper custom max pack mens running lab launch advance hurache retro classic training significant inspired breaking snkrs jordan importance futuristic brand heritage strength future cortez retailers special trancendant transformation court strike available athlete skate stength finesse innovate unlock sneakers"
+.split(' ');
+
+module.exports = words;
