+++ package.json
@@ -9,6 +9,7 @@
   },
   "scripts": {},
   "dependencies": {
+    "bluebird": "^3.7.2",
     "mysql": "^2.18.1"
   },
   "devDependencies": {
